+++ src/Symfony/Bundle/FrameworkBundle/Console/Application.php
@@ -13,7 +13,8 @@ namespace Symfony\Bundle\FrameworkBundle\Console;
 
 use Symfony\Component\Console\Application as BaseApplication;
 use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Command\ListCommand;
+use Symfony\Component\Console\Command\HelpCommand;
+use Symfony\Bundle\FrameworkBundle\Command\ListCommand;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\ConsoleOutputInterface;
@@ -217,5 +218,10 @@ class Application extends BaseApplication
                 $this->doRenderException($error, $output);
             }
         }
+    }
+
+    protected function getDefaultCommands()
+    {
+        return [new HelpCommand()];
     }
 }